test_uniform
